add copy_schema_role_grants_to_db_clones attribute to DatabaseBlueprint and update DatabaseParser so this config can be set from YAML
This attribute is intended to support schema role grant inheritance on
database clones. For example, if a database gets cloned from a source
db which is managed by SnowDDL, the grants on the source schema roles
will get duplexed over to the cloned db namespace. The benefit of this
is that one set of schema roles can be used to manage both the source
db and it's clones, and there is only minimal resource drift from the
clone operation (database level grants don't get copied)

This config is intended to be used with schema_roles: attribute in
SchemaBlueprint. If copying source schema role grants over to a cloned db,
presumably the user would not want additional schema roles generated for
that cloned db, so schema_roles should be set to False for the cloned db schemas